==> src/03_preprocessing/00_processor.ipynb <==
import pandas as pd
import random

import os, sys
sys.path.append(os.path.abspath("../../etc/"))
import config

df = pd.read_csv('../../data/CoV-AbDab_080224.csv')

def generate_hard_negatives(df, target_count, binding_col=config.BINDING_YES, neutral_col=config.NEUTRAL_YES):
    """
    Creates decoy negatives by shuffling CDR loops. 
    Maintains 'ND' and NaN values to preserve Nanobody vs Antibody logic.
    """
    
    # Sample from existing binders (Class 1) to ensure realistic scaffolds
    positives = df[df[binding_col] == config.TARGET].sample(n=target_count, replace=True if target_count > len(df) else False)
    negatives = positives.copy()

    def safe_shuffle(seq):
        # Return as is if it's ND, NaN, or too short to shuffle meaningfully
        if pd.isna(seq) or str(seq).upper() == 'ND' or len(str(seq)) < 2:
            return seq
        
        # Convert to list, shuffle, and rejoin
        char_list = list(str(seq))
        random.shuffle(char_list)
        return "".join(char_list)

    # Apply shuffling only to the CDR loops
    # Shuffling the 'fingers' but keeping the Framework intact
    negatives['CDRH3'] = negatives['CDRH3'].apply(safe_shuffle)
    
    if 'CDRL3' in negatives.columns:
        negatives['CDRL3'] = negatives['CDRL3'].apply(safe_shuffle)

    # Mark as non-binding and non-neutralizing
    negatives[config.NOT_BINDING] = config.TARGET
    negatives[config.NOT_NEUTRAL] = config.TARGET
    negatives[binding_col] = None
    negatives[neutral_col] = None
    
    # Optional: Tag them to keep track them in EDA
    negatives['data_source'] = 'synthetic_negative'
    
    return negatives


df_synthetic = generate_hard_negatives(df, target_count=5000)

# Combine with original data
df_balanced = pd.concat([df, df_synthetic], axis=0).reset_index(drop=True)

df_balanced.to_csv("./data/raw_transformation_00.csv")

%run "01_label_processor.ipynb"

%run "02_naive_processor.ipynb"

%run "03_motif_processor.ipynb"

%run "04_biochemical_processor.ipynb"
