==> src/03_preprocessing/01_label_processor.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import math
from enum import Enum
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import PCA
from itertools import product
from dotenv import load_dotenv

import os, sys
sys.path.append(os.path.abspath("../../etc/"))
import config

df_raw: pd.DataFrame = pd.read_csv('./data/raw_transformation_00.csv')
df_processed = pd.DataFrame()

RAW_FEATURES_SEQUENCE = config.EXTRACTABLE_BIOSEQUENCE_FEATURES

def data_cleaning(df: pd.DataFrame):
    df = df.drop(columns=config.IGNORED_FEATURES)
    has_existing_data = lambda df, col : df[df[col].notna() & (df[col]) != 'ND']
    for bio in RAW_FEATURES_SEQUENCE:
        df = has_existing_data(df, bio)
    return df

df_raw = data_cleaning(df_raw)

df_raw.head()

NEUTRAL_VS = config.NEUTRAL_YES or ""
NOT_NEUTRAL = config.NOT_NEUTRAL or ""
BINDING_VS = config.BINDING_YES or ""
NOT_BINDING = config.NOT_BINDING or ""

TARGET_NAME = config.TARGET or ""
NON_TARGETS_SARS_CoV2_WT = config.NON_TARGETS or ""


def label_neutralization(row, name: str):
   return 1 if one_hot_encode(row, NEUTRAL_VS, name) else (0 if one_hot_encode(row, NOT_NEUTRAL, name) else 2)

def label_binding(row, name: str):
    return 1 if one_hot_encode(row, BINDING_VS, name) else (0 if one_hot_encode(row, NOT_BINDING, name) else 2)

def label_antibody(row, name: str):
    return 1 if one_hot_encode(row, 'Ab or Nb', name) else 0
    
def get_list_from_cell(cell):
    """Safely converts semicolon-delimited strings to a set of strains."""
    if pd.isna(cell) or cell == '':
        return set()
    return set(str(cell).split(';'))

def one_hot_encode(row, col, name): 
    if pd.isna(row[col]) or row[col] == '':
        return False
    return name in str(row[col])

def label_binding_advanced(row):
    binds_set = get_list_from_cell(row[BINDING_VS])
    not_binds_set = get_list_from_cell(row[NOT_BINDING])
    
    # 1. PRIMARY RULE: Explicit Binder
    if TARGET_NAME in binds_set:
        return 1
    
    # 2. PRIMARY RULE: Explicit Non-Binder
    if TARGET_NAME in not_binds_set:
        return 0
    
    # 3. SECONDARY RULE: Hard Negatives
    # If it binds to MERS, OC43, etc., but NOT Wuhan, we treat it as 0
    if binds_set.intersection(NON_TARGETS_SARS_CoV2_WT):
        return 0
        
    return 2

df_processed[config.BINDING_TARGET] = df_raw.apply(label_binding_advanced, axis=1)
df_processed[config.NEUTRAL_TARGET] = df_raw.apply(lambda row: label_neutralization(row, TARGET_NAME), axis=1)
df_processed['is_nanobody'] = df_raw.apply(lambda row: label_antibody(row, 'Nb'), axis=1)
df_processed['name'] = df_raw['Name']

OUTFILE_PATH_LABELS = "./data/01_binding_labels.csv"
OUTFILE_PATH_RAW = "./data/raw_transformation_01.csv"

df_processed.to_csv(OUTFILE_PATH_LABELS, index=True)
df_raw.to_csv(OUTFILE_PATH_RAW, index=True)

df_processed['is_binding_SARS-CoV2_WT'].value_counts()

df_processed['is_neutral_SARS-CoV2_WT'].value_counts()

df_raw['Doesn\'t Bind to'].value_counts()
