==> src/03_preprocessing/02_naive_processor.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import math
from enum import Enum
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import PCA
from itertools import product

import os, sys
sys.path.append(os.path.abspath("../../etc/"))
import config

df_raw = pd.read_csv('./data/raw_transformation_01.csv', index_col=0)
df_processed = pd.read_csv('./data/01_binding_labels.csv', index_col=0)

df_raw.info()

get_length = lambda col : df_raw[col].str.len().fillna(0).astype(int)

lowercase_col_names = lambda col: str.lower(col).replace(' ', '_')

AMINO_ACIDS: str = config.AMINO_ACID_ALPHABETS or ""
BIOLOGICAL_FEATURES: list[str] = config.EXTRACTABLE_BIOSEQUENCE_FEATURES or []


def get_naive_biosequence_information(df: pd.DataFrame, df_processed: pd.DataFrame, col: str):
    print(f"Processing {col}...")
    
    temp_series = df[col].fillna("")
    lengths = temp_series.str.len()

    new_cols = {
        f'{col}_len': lengths.replace(0, 1) 
    }

    for aa in AMINO_ACIDS:
        counts = temp_series.str.count(aa)
        new_cols[f'{col}_amino_acid_percentage_{aa}'] = counts / new_cols[f'{col}_len']
        
    new_cols[f'{col}_len'] = lengths
    
    new_df = pd.concat([df_processed, pd.DataFrame(new_cols)], axis=1)
    return new_df

for seq in BIOLOGICAL_FEATURES:
    df_processed = get_naive_biosequence_information(df_raw, df_processed, seq)


df_processed.describe()

def get_one_hot_epitopes(df, column_name):
    clean_col = df[column_name].astype(str).str.upper()
    
    # Note: Order matters here if I want to avoid 'Other_Spike' overlapping
    mappings = {
        'S_RBD': clean_col.str.contains('RBD'),
        'S_NTD': clean_col.str.contains('NTD'),
        'S_S2':  clean_col.str.contains('S2'),
        'S_S1':  clean_col.str.contains('S1'),
        'N_Protein': clean_col.str.contains('N') & ~clean_col.str.contains('S')
    }
    encoding_df = pd.DataFrame(mappings).astype(int)
    is_spike = clean_col.str.contains('S')
    hit_any_specific = encoding_df.any(axis=1)
    encoding_df['Other_Spike'] = (is_spike & ~hit_any_specific).astype(int)
    encoding_df['Unknown'] = (df[column_name].isna() | 
                              clean_col.isin(['UNKNOWN', 'TBC', 'NAN']) | 
                              ~encoding_df.any(axis=1)).astype(int)
    
    return encoding_df

# Usage:
one_hot_df = get_one_hot_epitopes(df_raw, 'Protein + Epitope')

df_processed = pd.concat([df_processed, one_hot_df], axis=1)

OUTFILE_PATH = './data/02_naive_processed_features.csv'
df_processed.drop(columns=[config.BINDING_TARGET,config.NEUTRAL_TARGET,config.IS_NANOBODY_COL,'name']).to_csv(OUTFILE_PATH)

df_processed
