==> src/03_preprocessing/03_motif_processor.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import math
from enum import Enum
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import PCA
from itertools import product
import sys
from pathlib import Path

import os, sys
sys.path.append(os.path.abspath("../../etc/"))
import config

df_raw = pd.read_csv('./data/raw_transformation_01.csv', index_col=0)
df_processed = pd.read_csv('./data/01_binding_labels.csv', index_col=0)

biological_sequence_list = config.EXTRACTABLE_BIOSEQUENCE_FEATURES

get_length = lambda col: df_raw[col].str.len().fillna(0).astype(int)
lowercase_col_names = lambda col: str.lower(col).replace(' ', '_')

AMINO_ACIDS = config.AMINO_ACID_ALPHABETS

CONJOINT_TRIADS = {
    'A': '1', 'G': '1', 'V': '1',             # Aliphatic
    'I': '2', 'L': '2', 'F': '2', 'P': '2',   # Large Aliphatic
    'Y': '3', 'M': '3', 'T': '3', 'S': '3',   # Polar
    'H': '4', 'N': '4', 'Q': '4', 'W': '4',   # Aromatic/Neutral
    'R': '5', 'K': '5',                       # Positive Charge
    'D': '6', 'E': '6',                       # Negative Charge
    'C': '7'                                  # Cysteine (Special)
}

#Pre-generate the list of all 343 possible triads
TRIAD_NAMES = ["".join(t) for t in product("1234567", repeat=3)]
KMER3_NAMES = ["".join(t) for t in product(config.AMINO_ACID_ALPHABETS, repeat=3)]

def get_conjoint_triads_fast(df, col):
    # Vectorized translation of sequence to groups
    translation_table = str.maketrans(CONJOINT_TRIADS)
    # We remove non-standard AAs by ignoring anything not in our 7 groups
    encoded_series = df[col].fillna("").str.translate(translation_table)
    
    # Prepare the CountVectorizer for 3-character "words"
    # analyzer='char' and ngram_range=(3,3) mimics the triad sliding window
    
    vectorizer = CountVectorizer(analyzer='char', ngram_range=(3, 3), vocabulary=TRIAD_NAMES)
    triad_matrix = vectorizer.transform(encoded_series)
    lengths = encoded_series.str.len() - 2
    lengths = lengths.clip(lower=1) # Prevent div by zero
    triad_df = pd.DataFrame(
        triad_matrix.toarray() / lengths.values[:, None], # type: ignore
        columns=[f"{col}_CTD_{t}" for t in TRIAD_NAMES],
        index=df.index
    )
    
    return triad_df



def get_3kmer_motif_triads_fast(df: pd.DataFrame, col: str):
    # We remove non-standard AAs by ignoring anything not in our 7 groups
    encoded_series = df[col].fillna("")
    
    # Prepare the CountVectorizer for 3-character "words"
    # analyzer='char' and ngram_range=(3,3) mimics the triad sliding window
    vectorizer = CountVectorizer(analyzer='char', ngram_range=(3, 3), vocabulary=KMER3_NAMES, lowercase=False)
    triad_matrix = vectorizer.transform(encoded_series)
    lengths = encoded_series.str.len() - 2
    lengths = lengths.clip(lower=1) # Prevent div by zero
    
    triad_df = pd.DataFrame(
        triad_matrix.toarray() / lengths.values[:, None], # type: ignore
        columns=[f"{col}_3kmer_{t}" for t in KMER3_NAMES],
        index=df.index,
        
    )
    
    return triad_df

feature_conjoint_triads_dfs = []
feature_kmer_3_dfs = []


for seq in biological_sequence_list:
    # Get CTD info as its own DF
    ctd_df = get_conjoint_triads_fast(df_raw, seq)
    feature_conjoint_triads_dfs.append(ctd_df)
    
    # kmer3_df = get_3kmer_motif_triads_fast(df_raw, seq)
    # feature_kmer_3_dfs.append(kmer3_df)
    
df_processed_conjoint = pd.concat([df_processed] + feature_conjoint_triads_dfs, axis=1)
# df_processed_motif = pd.concat([df_processed] + feature_kmer_3_dfs, axis=1)

df_processed_conjoint.describe()

# df_processed_motif.describe()

OUTFILE_PATH_CONJOINT = './data/03_conjoint_triads_processed_features.csv'
# OUTFILE_PATH_MOTIF = './data/03_kmer3_processed_features.csv'
df_processed_conjoint.drop(columns=[config.BINDING_TARGET,config.NEUTRAL_TARGET,config.IS_NANOBODY_COL]).to_csv(OUTFILE_PATH_CONJOINT)
# df_processed_motif.drop(columns=[config.BINDING_TARGET,config.NEUTRAL_TARGET,config.IS_NANOBODY_COL]).to_csv(OUTFILE_PATH_MOTIF)
