==> src/03_preprocessing/04_biochemical_processor.ipynb <==
import pandas as pd
from peptides import Peptide
from Bio.SeqUtils.ProtParam import ProteinAnalysis
import re
import os, sys
sys.path.append(os.path.abspath("../../etc/"))
import config

df_raw = pd.read_csv("./data/raw_transformation_01.csv")
df_processed = pd.read_csv("./data/01_binding_labels.csv")

def extract_biochemical_features(sequence, col):
    # Handle NaNs or empty strings
    if pd.isna(sequence) or str(sequence).strip() == "":
        return None
    
    # SANITIZATION: Remove spaces and non-AA characters
    # This keeps only A-Z and removes anything else (like spaces or '*')
    clean_seq = re.sub(r'[^A-Z]', '', str(sequence).upper())
    
    # Biopython's instability_index requires at least 2 AAs
    if len(clean_seq) < 2:
        return None

    try:
        p = Peptide(clean_seq)
        analysed_seq = ProteinAnalysis(clean_seq)
        
        features = {
            # 1. Physicochemical
            f"{col}_hydrophobicity": p.hydrophobicity(),
            f"{col}_isoelectric_point": p.isoelectric_point(),
            f"{col}_charge_ph7": p.charge(pH=7.4),
            f"{col}_molecular_weight": p.molecular_weight(),
            
            # 2. Stability & Structure
            f"{col}_aliphatic_index": p.aliphatic_index(),
            f"{col}_instability_index": analysed_seq.instability_index(),
            f"{col}_aromaticity": analysed_seq.aromaticity(),
            
            # 3. Binding Potential
            f"{col}_boman_index": p.boman(),
            f"{col}_flexibility": sum(analysed_seq.flexibility()) / len(clean_seq)
        }
        return features
    except KeyError as e:
        
        print(f"Skipping sequence {clean_seq} due to invalid character: {e}")
        return None



feature_df = []
for seq in config.EXTRACTABLE_BIOSEQUENCE_FEATURES:
    feature_df.append(df_raw[seq].apply(lambda row: extract_biochemical_features(row, seq)).apply(pd.Series))

out_df = pd.concat(feature_df,axis=1)

OUTFILE_PATH = "./data/04_antibody_only_physicochemical_properties.csv"
out_df.to_csv(OUTFILE_PATH)

out_df
